add process dump after failed install
